+++ RELEASES.md
@@ -1541,3 +1541,13 @@ Based on:
 - [typescript v0.1.0-alpha.148] .
 ### Releases
 - [NPM v0.1.0-alpha.148] https://www.npmjs.com/package/syllable-sdk/v/0.1.0-alpha.148 - .
+
+## 2025-05-14 14:43:47
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.493.2 (2.512.4) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v0.1.0-alpha.149] .
+### Releases
+- [NPM v0.1.0-alpha.149] https://www.npmjs.com/package/syllable-sdk/v/0.1.0-alpha.149 - .